Change where the state lives.
